requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
joblib==1.1.0
Markdown==3.3.4
numpy==1.21.4
pycparser==2.21
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
threadpoolctl==3.0.0
